# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/expenditure": {
            "get": {
                "tags": [
                    "Expenditures"
                ],
                "summary": "View an Expenditure",
                "operationId": "7da058a1a678fdef1ef32c33bfb39658",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "value": {
                                            "type": "string|decimal"
                                        },
                                        "data_cadastro": {
                                            "type": "string|dateTime('Y-m-d')"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "data": {
                                                "id": 1,
                                                "description": "Creatina",
                                                "value": "28.89",
                                                "data_cadastro": "2024-01-12"
                                            },
                                            "0": {
                                                "id": 1,
                                                "description": "Creatina",
                                                "value": "28.89",
                                                "data_cadastro": "2024-01-12"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenditures"
                ],
                "summary": "Create an Expenditure",
                "operationId": "ac72fc23720b866b76a3347a994b417e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "required|string|max:191"
                                    },
                                    "value": {
                                        "type": "required|decimal:2|between:0,99999999.99"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "Creatina 300gr",
                                    "value": 123456.78
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "data": {
                                                "message": "Despesa cadastrada com sucesso!"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/expenditure/{id}": {
            "get": {
                "tags": [
                    "Expenditures"
                ],
                "summary": "View an Expenditure",
                "operationId": "190f95b2f22be7199d3231a080eacdc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expenditue Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "value": {
                                            "type": "string|decimal"
                                        },
                                        "data_cadastro": {
                                            "type": "string|dateTime('Y-m-d')"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "data": {
                                                "id": 1,
                                                "description": "Creatina",
                                                "value": "28.89",
                                                "data_cadastro": "2024-01-12"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenditures"
                ],
                "summary": "Updates a Expenditure",
                "operationId": "fed704b1173da4b7d9c7fc179e005c24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expenditue Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "required|string|max:191"
                                    },
                                    "value": {
                                        "type": "required|decimal:2|between:0,99999999.99"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "Creatina 300gr",
                                    "value": 123456.78
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "data": {
                                                "message": "Despesa atualizada com sucesso!"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expenditures"
                ],
                "summary": "Delete an Expenditure",
                "operationId": "501190427cfa49f4dd99665e99e94f55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expenditue Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "data": {
                                                "message": "Despesa removida com sucesso!"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}